Update usage in API client to validate params before concatenating

+++ includes/wc-payment-api/class-wc-payments-api-client.php
@@ -231,8 +231,17 @@ class WC_Payments_API_Client {
 	 * @param string $intent_id intent id.
 	 *
 	 * @return WC_Payments_API_Payment_Intention intention object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_intent( $intent_id ) {
+		if ( ! preg_match( '/\w+/', $intent_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$intent = $this->request( [], self::INTENTIONS_API . '/' . $intent_id, self::GET );
 
 		return $this->deserialize_payment_intention_object_from_array( $intent );
@@ -448,8 +457,17 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $transaction_id id of requested transaction.
 	 * @return array transaction object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_transaction( $transaction_id ) {
+		if ( ! preg_match( '/\w+/', $transaction_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$transaction = $this->request( [], self::TRANSACTIONS_API . '/' . $transaction_id, self::GET );
 
 		if ( is_wp_error( $transaction ) ) {
@@ -521,7 +539,7 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function get_dispute( $dispute_id ) {
-		if ( ! preg_match( '/[A-Za-z0-9_-]+/', $dispute_id ) ) {
+		if ( ! preg_match( '/\w+/', $dispute_id ) ) {
 			throw new API_Exception(
 				__( 'Route param validation failed.', 'woocommerce-payments' ),
 				'wcpay_route_validation_failure',
@@ -548,8 +566,17 @@ class WC_Payments_API_Client {
 	 * @param array  $metadata   metadata associated with this dispute.
 	 *
 	 * @return array dispute object.
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function update_dispute( $dispute_id, $evidence, $submit, $metadata ) {
+		if ( ! preg_match( '/\w+/', $dispute_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$request = [
 			'evidence' => $evidence,
 			'submit'   => $submit,
@@ -574,8 +601,18 @@ class WC_Payments_API_Client {
 	 *
 	 * @param string $dispute_id id of dispute to close.
 	 * @return array dispute object.
+	 *
+	 * @throws API_Exception - Exception thrown in case route validation fails.
 	 */
 	public function close_dispute( $dispute_id ) {
+		if ( ! preg_match( '/\w+/', $dispute_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		$dispute = $this->request( [], self::DISPUTES_API . '/' . $dispute_id . '/close', self::POST );
 		// Invalidate the dispute caches.
 		\WC_Payments::get_database_cache()->delete( Database_Cache::DISPUTE_STATUS_COUNTS_KEY );
@@ -695,6 +732,14 @@ class WC_Payments_API_Client {
 	 */
 	public function get_file_contents( string $file_id, bool $as_account = true ): array {
 		try {
+			if ( ! preg_match( '/\w+/', $file_id ) ) {
+				throw new API_Exception(
+					__( 'Route param validation failed.', 'woocommerce-payments' ),
+					'wcpay_route_validation_failure',
+					400
+				);
+			}
+
 			return $this->request( [ 'as_account' => $as_account ], self::FILES_API . '/' . $file_id . '/contents', self::GET );
 		} catch ( API_Exception $e ) {
 			Logger::error( 'Error retrieving file contents for ' . $file_id . '. ' . $e->getMessage() );
@@ -712,6 +757,13 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function get_file( string $file_id, bool $as_account = true ): array {
+		if ( ! preg_match( '/\w+/', $file_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
 		return $this->request( [ 'as_account' => $as_account ], self::FILES_API . '/' . $file_id, self::GET );
 	}
 
@@ -1118,6 +1170,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If fetching the invoice fails.
 	 */
 	public function get_invoice( string $invoice_id ) {
+		if ( ! preg_match( '/\w+/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::INVOICES_API . '/' . $invoice_id,
@@ -1137,6 +1197,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception Error charging the invoice.
 	 */
 	public function charge_invoice( string $invoice_id, array $data = [] ) {
+		if ( ! preg_match( '/\w+/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::INVOICES_API . '/' . $invoice_id . '/pay',
@@ -1154,6 +1222,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception Error updating the invoice.
 	 */
 	public function update_invoice( string $invoice_id, array $data = [] ) {
+		if ( ! preg_match( '/\w+/', $invoice_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::INVOICES_API . '/' . $invoice_id,
@@ -1171,6 +1247,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function update_charge( string $charge_id, array $data = [] ) {
+		if ( ! preg_match( '/\w+/', $charge_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::CHARGES_API . '/' . $charge_id,
@@ -1187,7 +1271,7 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function get_charge( string $charge_id ) {
-		if ( ! preg_match( '/[A-Za-z0-9_-]+/', $charge_id ) ) {
+		if ( ! preg_match( '/\w+/', $charge_id ) ) {
 			throw new API_Exception(
 				__( 'Route param validation failed.', 'woocommerce-payments' ),
 				'wcpay_route_validation_failure',
@@ -1212,6 +1296,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception
 	 */
 	public function update_transaction( string $transaction_id, array $data = [] ) {
+		if ( ! preg_match( '/\w+/', $transaction_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::TRANSACTIONS_API . '/' . $transaction_id,
@@ -1229,6 +1321,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If fetching the subscription fails.
 	 */
 	public function get_subscription( string $wcpay_subscription_id ) {
+		if ( ! preg_match( '/\w+/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1264,6 +1364,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If updating the WCPay subscription fails.
 	 */
 	public function update_subscription( $wcpay_subscription_id, $data ) {
+		if ( ! preg_match( '/\w+/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1281,6 +1389,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If canceling the subscription fails.
 	 */
 	public function cancel_subscription( string $wcpay_subscription_id ) {
+		if ( ! preg_match( '/\w+/', $wcpay_subscription_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::SUBSCRIPTIONS_API . '/' . $wcpay_subscription_id,
@@ -1299,6 +1415,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If updating the WCPay subscription item fails.
 	 */
 	public function update_subscription_item( $wcpay_subscription_item_id, $data ) {
+		if ( ! preg_match( '/\w+/', $wcpay_subscription_item_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$data,
 			self::SUBSCRIPTION_ITEMS_API . '/' . $wcpay_subscription_item_id,
@@ -1316,6 +1440,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If payment method does not exist.
 	 */
 	public function get_payment_method( $payment_method_id ) {
+		if ( ! preg_match( '/\w+/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id,
@@ -1334,6 +1466,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If payment method update fails.
 	 */
 	public function update_payment_method( $payment_method_id, $payment_method_data = [] ) {
+		if ( ! preg_match( '/\w+/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			$payment_method_data,
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id,
@@ -1353,6 +1493,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If an error occurs.
 	 */
 	public function get_payment_methods( $customer_id, $type, $limit = 100 ) {
+		if ( ! preg_match( '/\w+/', $customer_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[
 				'customer' => $customer_id,
@@ -1374,6 +1522,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If detachment fails.
 	 */
 	public function detach_payment_method( $payment_method_id ) {
+		if ( ! preg_match( '/\w+/', $payment_method_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request(
 			[],
 			self::PAYMENT_METHODS_API . '/' . $payment_method_id . '/detach',
@@ -1539,6 +1695,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If an error occurs.
 	 */
 	public function update_terminal_location( $location_id, $display_name, $address ) {
+		if ( ! preg_match( '/\w+/', $location_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		// Any parameters not provided will be left unchanged so pass only supplied values.
 		$update_request_body = array_merge(
 			( isset( $address ) ? [ 'address' => $address ] : [] ),
@@ -1563,6 +1727,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception If the location id is invalid or downstream call fails.
 	 */
 	public function delete_terminal_location( $location_id ) {
+		if ( ! preg_match( '/\w+/', $location_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request( [], self::TERMINAL_LOCATIONS_API . '/' . $location_id, self::DELETE );
 	}
 
@@ -1599,6 +1771,14 @@ class WC_Payments_API_Client {
 	 * @throws API_Exception - If not connected or request failed.
 	 */
 	public function get_document( $document_id ) {
+		if ( ! preg_match( '/\w+/', $document_id ) ) {
+			throw new API_Exception(
+				__( 'Route param validation failed.', 'woocommerce-payments' ),
+				'wcpay_route_validation_failure',
+				400
+			);
+		}
+
 		return $this->request( [], self::DOCUMENTS_API . '/' . $document_id, self::GET, true, false, true );
 	}
